장애물 enum 개선, 매직넘버 제거 #3
-  장애물 매직넘버 제거
-  장애물 enum class 미적용분 추가

+++ JustJump/src/core/GameWorld.cpp
@@ -298,7 +298,8 @@ void GameWorld::OnChar(WPARAM ch)
 	{
         int offset = GameConst::kPortalCheatOffset;
 		for (const auto& obs : m_object_manager.GetObjects()) {
-			if (obs->GetType() == static_cast<int>(EObstacleType::Portal)) {
+			if (obs->GetType() == EObstacleType::Portal)
+            {
 				m_player.SetX(obs->GetX() + offset);
 				m_player.SetY(obs->GetY() + offset); // 발 위치 맞추기
 				break;
